[extensionservices] Migrate to bnd
