--- addons/extensionservice/org.openhab.extensionservice.marketplace.automation/META-INF/MANIFEST.MF
@@ -1,25 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ActivationPolicy: lazy
-Bundle-ClassPath: .
-Bundle-ManifestVersion: 2
-Bundle-Name: IoT Marketplace Extension Automation Supp
- ort
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-SymbolicName: org.openhab.extensionservice.marketplace.
- automation;singleton:=true
-Bundle-Vendor: openHAB
-Bundle-Version: 2.5.0.qualifier
-Import-Package: 
- org.apache.commons.io,
- org.eclipse.jdt.annotation;resolution:=optional,
- org.openhab.core.automation.parser,
- org.openhab.core.automation.template,
- org.eclipse.smarthome.config.core,
- org.eclipse.smarthome.core.common.registry,
- org.eclipse.smarthome.core.extension,
- org.eclipse.smarthome.core.storage,
- org.openhab.extensionservice.marketplace,
- org.osgi.framework,
- org.slf4j
-Service-Component: OSGI-INF/*.xml
-Automatic-Module-Name: org.openhab.extensionservice.marketplace.automation
