--- addons/extensionservice/org.openhab.extensionservice.marketplace.automation/OSGI-INF/.gitignore
@@ -1,1 +1,0 @@
-/*.xml
