--- addons/extensionservice/org.openhab.extensionservice.marketplace.automation/build.properties
@@ -1,6 +1,0 @@
-output.. = target/classes
-bin.includes = META-INF/,\
-               .,\
-               OSGI-INF/,\
-               NOTICE
-source.. = src/main/java/
