--- addons/extensionservice/org.openhab.extensionservice.marketplace.automation/pom.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.openhab.extensionservice</groupId>
-    <artifactId>pom</artifactId>
-    <version>2.5.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>org.openhab.extensionservice.marketplace.automation</artifactId>
-  <packaging>eclipse-plugin</packaging>
-
-  <name>IoT Marketplace Extension Automation Support</name>
-
-</project>
