--- addons/extensionservice/org.openhab.extensionservice.marketplace.test/build.properties
@@ -1,6 +1,0 @@
-source.. = src/test/java/
-output.. = target/test-classes/
-bin.includes = META-INF/,\
-               .,\
-               NOTICE
-
