--- addons/extensionservice/org.openhab.extensionservice.marketplace/OSGI-INF/.gitignore
@@ -1,1 +1,0 @@
-/*.xml
