--- addons/extensionservice/org.openhab.extensionservice.marketplace/build.properties
@@ -1,7 +1,0 @@
-output.. = target/classes
-bin.includes = META-INF/,\
-               .,\
-               OSGI-INF/,\
-               NOTICE,\
-               ESH-INF/
-source.. = src/main/java/
