--- addons/extensionservice/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.openhab.addons</groupId>
-    <artifactId>pom</artifactId>
-    <version>2.5.0-SNAPSHOT</version>
-  </parent>
-
-  <groupId>org.openhab.extensionservice</groupId>
-  <artifactId>pom</artifactId>
-
-  <packaging>pom</packaging>
-
-  <name>Extension Service Extensions</name>
-
-  <modules>
-    <module>org.openhab.extensionservice.marketplace</module>
-    <module>org.openhab.extensionservice.marketplace.automation</module>
-    <module>org.openhab.extensionservice.marketplace.test</module>
-  </modules>
-
-</project>
