+++ bundles/org.openhab.extensionservice.marketplace.automation/NOTICE
