+++ bundles/org.openhab.extensionservice.marketplace.automation/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>org.openhab.addons.bundles</groupId>
+		<artifactId>org.openhab.addons.reactor.bundles</artifactId>
+		<version>2.5.0-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>org.openhab.extensionservice.marketplace.automation</artifactId>
+
+	<name>openHAB Add-ons :: Extension Service :: Eclipse IoT Marketplace (Automation)</name>
+
+	<dependencies>
+		<dependency>
+			<groupId>org.openhab.addons.bundles</groupId>
+			<artifactId>org.openhab.extensionservice.marketplace</artifactId>
+			<version>${project.version}</version>
+			<scope>compile</scope>
+		</dependency>
+	</dependencies>
+</project>
