+++ bundles/org.openhab.extensionservice.marketplace/src/main/resources/ESH-INF/config/config.xml
